--- a/augly/video/augmenters/ffmpeg/aspect_ratio.py
+++ b/augly/video/augmenters/ffmpeg/aspect_ratio.py
@@ -2,32 +2,31 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 
 import math
-from typing import Dict, Tuple, Union
+from typing import List, Union
 
-from augly.video.augmenters.ffmpeg.base_augmenter import BaseFFMPEGAugmenter
+from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter
 from augly.video.helpers import get_video_info
-from ffmpeg.nodes import FilterableStream
 
 
-class VideoAugmenterByAspectRatio(BaseFFMPEGAugmenter):
+class VideoAugmenterByAspectRatio(BaseVidgearFFMPEGAugmenter):
     def __init__(self, ratio: Union[float, str]):
         assert (isinstance(ratio, str) and len(ratio.split(":")) == 2) or (
             isinstance(ratio, (int, float)) and ratio > 0
         ), "Aspect ratio must be a valid string ratio or a positive number"
         self.aspect_ratio = ratio
 
-    def add_augmenter(
-        self, in_stream: FilterableStream, **kwargs
-    ) -> Tuple[FilterableStream, Dict]:
+    def get_command(self, video_path: str, output_path: str) -> List[str]:
         """
         Changes the sample (sar) & display (dar) aspect ratios of the video
 
-        @param in_stream: the FFMPEG object of the video
+        @param video_path: the path to the video to be augmented
 
-        @returns: a tuple containing the FFMPEG object with the augmentation
-            applied and a dictionary with any output arguments as necessary
+        @param output_path: the path in which the resulting video will be stored.
+
+        @returns: a list of strings of the FFMPEG command if it were to be written
+            in a command line
         """
-        video_info = get_video_info(kwargs["video_path"])
+        video_info = get_video_info(video_path)
 
         area = int(video_info["width"]) * int(video_info["height"])
         if isinstance(self.aspect_ratio, float):
@@ -39,10 +38,19 @@ class VideoAugmenterByAspectRatio(BaseFFMPEGAugmenter):
         new_w = int(math.sqrt(area * aspect_ratio))
         new_h = int(area / new_w)
 
-        return (
-            in_stream.video.filter("scale", **{"width": new_w, "height": new_h})
-            .filter("pad", **{"width": "ceil(iw/2)*2", "height": "ceil(ih/2)*2"})
-            .filter("setsar", **{"ratio": self.aspect_ratio})
-            .filter("setdar", **{"ratio": self.aspect_ratio}),
-            {},
-        )
+        command = [
+            "-y",
+            "-i",
+            video_path,
+            "-vf",
+            f"scale=width={new_w}:height={new_h},"
+            + f"pad=width=ceil(iw/2)*2:height=ceil(ih/2)*2,"
+            + f"setsar=ratio={self.aspect_ratio},"
+            + f"setdar=ratio={self.aspect_ratio}",
+            "-c:a",
+            "copy",
+            "-preset",
+            "ultrafast",
+            output_path,
+        ]
+        return command