--- a/augly/video/augmenters/ffmpeg/aspect_ratio.py
+++ b/augly/video/augmenters/ffmpeg/aspect_ratio.py
@@ -53,5 +53,5 @@ class VideoAugmenterByAspectRatio(BaseVidgearFFMPEGAugmenter):
             "ultrafast",
             output_path,
         ]
-        
+
         return command