--- a/augly/video/augmenters/ffmpeg/audio_swap.py
+++ b/augly/video/augmenters/ffmpeg/audio_swap.py
@@ -23,8 +23,8 @@ class VideoAugmenterByAudioSwap(BaseVidgearFFMPEGAugmenter):
 
         @param output_path: the path in which the resulting video will be stored.
 
-        @returns: a list of strings of the FFMPEG command if it were to be written
-            in a command line
+        @returns: a list of strings containing the CLI FFMPEG command for
+            the augmentation
         """
         audio_info = get_audio_info(self.audio_path)
         video_info = get_video_info(video_path)