--- a/augly/video/functional.py
+++ b/augly/video/functional.py
@@ -129,9 +129,8 @@ def audio_swap(
     """
     func_kwargs = helpers.get_func_kwargs(metadata, locals(), video_path)
 
-    af.VideoAugmenterByAudioSwap(audio_path, offset).add_augmenter(
-        video_path, output_path
-    )
+    audio_swap_aug = af.VideoAugmenterByAudioSwap(audio_path, offset)
+    audio_swap_aug.add_augmenter(video_path, output_path)
 
     if metadata is not None:
         helpers.get_metadata(
@@ -362,7 +361,8 @@ def change_aspect_ratio(
     """
     func_kwargs = helpers.get_func_kwargs(metadata, locals(), video_path)
 
-    af.VideoAugmenterByAspectRatio(ratio).add_augmenter(video_path, output_path)
+    aspect_ratio_aug = af.VideoAugmenterByAspectRatio(ratio)
+    aspect_ratio_aug.add_augmenter(video_path, output_path)
 
     if metadata is not None:
         helpers.get_metadata(
@@ -1163,9 +1163,10 @@ def overlay(
         else:
             resize(overlay_path, tmp_overlay_path, overlay_h, overlay_w)
 
-    af.VideoAugmenterByOverlay(
+    overlay_aug = af.VideoAugmenterByOverlay(
         tmp_overlay_path or overlay_path, x_factor, y_factor, use_overlay_audio
-    ).add_augmenter(video_path, output_path)
+    )
+    overlay_aug.add_augmenter(video_path, output_path)
 
     if tmp_overlay_path:
         os.remove(tmp_overlay_path)